+++ CHANGELOG.rst
@@ -37,6 +37,7 @@ TODO: integration test with pandapower (see `pandapower/contingency/contingency.
 - [IMPROVED] clean separation between lightsim2grid_core the main library that could be
   used from cpp and the python bindings.
 - [IMPROVED] removed the ".values" and replace them by ".to_numpy()" in pandapower converter.
+- [IMPROVED] remove the "typedef" in favor of "using" cpp side (core)
 
 [0.13.1]  2026-04-21
 --------------------